Planner backend test
Just tests that signon is hooked up, the app loads and booking requests
are viewed, thus the locations API is also accessible.

I added configuration for the staging environment for completion-sake
but it's likely we'll remove the staging configuration across all
checks soon, since we get lots of false-negatives due to the time it
takes to spin up heroku dynos in the staging environments.